feat: add rdbms exporter + batch operation execute applier + adjust events

+++ search/search-domain/src/main/java/io/camunda/search/entities/BatchOperationEntity.java
@@ -16,7 +16,7 @@ public record BatchOperationEntity(
     String operationType,
     OffsetDateTime startDate,
     OffsetDateTime endDate,
-    Long operationsTotalCount,
-    Long operationsFailedCount,
-    Long operationsCompletedCount) {
+    Integer operationsTotalCount,
+    Integer operationsFailedCount,
+    Integer operationsCompletedCount) {
 }
